3단계. LottoBonusNumberValidator 분리
