+++ Lotto/Lotto/Lotto/Model/LottoBonusNumberValidator.swift
@@ -1,0 +1,19 @@
+//
+//  LottoBonusNumberValidator.swift
+//  Lotto
+//
+//  Created by Kang, Su Jin (강수진) on 2022/05/17.
+//
+
+import Foundation
+
+struct LottoBonusNumberValidator {
+    
+    private let lottoNumbersValidator = LottoNumbersValidator()
+    
+    func validate(_ bonusNumber: Int, in winningLotto: Lotto) throws {
+        try lottoNumbersValidator.validateNumberInRange(of: bonusNumber)
+        let winningNumbersWithBonus = winningLotto.numbers + [bonusNumber]
+        try lottoNumbersValidator.validateAllUniqueNumbers(of: winningNumbersWithBonus)
+    }
+}
