+++ Lotto/LottoTest/Lotto/LottoNumbersValidatorTests.swift
@@ -89,45 +89,6 @@ class LottoNumbersValidatorTests: XCTestCase {
         }
     }
     
-    // MARK: - validateBonusNumber
-    
-    func test_validateBonusNumber() throws {
-        //given
-        let bonusNumber = 45
-        let winningNumbers = [1,2,3,4,5,6]
-        let winningLotto = try Lotto(numbers: winningNumbers)
-        
-        // when
-        // then
-        XCTAssertNoThrow(try sut.validateBonusNumber(bonusNumber, in: winningLotto))
-    }
-    
-    func test_validateBonusNumber_whenBonusNumbersIsDuplicated_throwHasDuplicate() throws {
-        //given
-        let bonusNumber = 1
-        let winningNumbers = [1,2,3,4,5,6]
-        let winningLotto = try Lotto(numbers: winningNumbers)
-        
-        // then
-        let expectation = LottoNumbersValidator.LottoNumbersValidatorError.hasDuplicate
-        XCTAssertThrowsError(try sut.validateBonusNumber(bonusNumber, in: winningLotto)) { error in
-            XCTAssertEqual(error as? LottoNumbersValidator.LottoNumbersValidatorError, expectation)
-        }
-    }
-    
-    func test_validateBonusNumber_whenBonusNumbersIsOverRanged_throwContainOutOfRange() throws {
-        //given
-        let bonusNumber = 46
-        let winningNumbers = [1,2,3,4,5,6]
-        let winningLotto = try Lotto(numbers: winningNumbers)
-        
-        // then
-        let expectation = LottoNumbersValidator.LottoNumbersValidatorError.containOutOfRange
-        XCTAssertThrowsError(try sut.validateBonusNumber(bonusNumber, in: winningLotto)) { error in
-            XCTAssertEqual(error as? LottoNumbersValidator.LottoNumbersValidatorError, expectation)
-        }
-    }
-    
     // MARK: - guideDescription
     
     func test_guideDescription_whenNumbersCountIsUnderSix() {
